check if config file in theme exists

+++ vc-platform-installer/install-theme.py
@@ -41,10 +41,11 @@ def copy_content(from_path, to_path):
     if(os.path.exists(from_folder)):
         copy_tree(from_folder, to_folder)
         print(datetime.datetime.now())
-        print('Copying complete')
     settingsFileFrom = os.path.join(from_path, settingsPath)
     settingsFileTo = os.path.join(to_path, settingsPath)
-    copy_file(settingsFileFrom, settingsFileTo)
+    if(os.path.exists(settingsFileFrom)):
+        copy_file(settingsFileFrom, settingsFileTo)
+    print('Copying complete')
 
 def removing_old_theme(path):
     print(datetime.datetime.now())
